Refactor the Postings struct into modules implementing the separate concerns.
