--- src/series/index/postings.rs
@@ -1,1985 +1,0 @@
-use super::index_key::IndexKey;
-use super::key_buffer::KeyBuffer;
-use crate::common::logging::log_warning;
-use crate::error_consts::{INTERNAL_ERROR, MISSING_FILTER};
-use crate::labels::filters::{
-    FilterList, LabelFilter, MatchOp, PredicateMatch, PredicateValue, SeriesSelector,
-};
-use crate::labels::{InternedLabel, SeriesLabel};
-use crate::series::{SeriesRef, TimeSeries};
-use blart::map::Entry as ARTEntry;
-use blart::{AsBytes, TreeMap};
-use croaring::Bitmap64;
-use smallvec::SmallVec;
-use std::borrow::Cow;
-use std::cmp::Ordering;
-use std::collections::{BTreeMap, BTreeSet};
-use std::ops::Bound;
-use std::sync::LazyLock;
-use valkey_module::{ValkeyError, ValkeyResult};
-
-pub(super) static EMPTY_BITMAP: LazyLock<PostingsBitmap> = LazyLock::new(PostingsBitmap::new);
-
-pub type PostingsBitmap = Bitmap64;
-// label
-// label=value
-pub type PostingsIndex = TreeMap<IndexKey, PostingsBitmap>;
-
-/// Type for the key of the index.
-pub type KeyType = Box<[u8]>;
-
-/// One series buffered during a load window, consumed by [`Postings::bulk_index`]
-pub(crate) struct BulkIndexEntry {
-    pub id: SeriesRef,
-    pub key: KeyType,
-    /// Precomputed `label=value` index keys for every label of the series.
-    pub label_keys: Vec<IndexKey>,
-}
-
-/// `Postings` is the core in-memory inverted index for time series data. It is designed for efficient
-/// querying and retrieving of time series based on their labels.
-#[derive(Clone)]
-pub struct Postings {
-    /// Map from label name and (label name, label value) to a set of timeseries ids.
-    pub(super) label_index: PostingsIndex,
-    /// Map from timeseries id to the key of the timeseries.
-    pub(in crate::series) id_to_key: BTreeMap<SeriesRef, KeyType>,
-    /// Set of timeseries ids of series that should be removed from the index. This really only
-    /// happens when the index is inconsistent (value does not exist in the db but exists in the index)
-    /// Keep track and cleanup from the index during a gc pass.
-    pub(super) stale_ids: PostingsBitmap,
-    /// Set of all timeseries ids in the index. This is used to optimize queries that are subtractive.
-    pub(crate) all_postings: PostingsBitmap,
-}
-
-impl Default for Postings {
-    fn default() -> Self {
-        Postings {
-            label_index: PostingsIndex::new(),
-            id_to_key: BTreeMap::new(),
-            stale_ids: PostingsBitmap::default(),
-            all_postings: PostingsBitmap::default(),
-        }
-    }
-}
-
-impl Postings {
-    #[allow(dead_code)]
-    pub(super) fn clear(&mut self) {
-        self.label_index.clear();
-        self.id_to_key.clear();
-        self.stale_ids.clear();
-        self.all_postings.clear();
-    }
-
-    /// `swap` the inner value with some other value
-    /// this is specifically to handle the `swapdb` event callback
-    pub fn swap(&mut self, other: &mut Self) {
-        std::mem::swap(&mut self.label_index, &mut other.label_index);
-        std::mem::swap(&mut self.id_to_key, &mut other.id_to_key);
-        std::mem::swap(&mut self.stale_ids, &mut other.stale_ids);
-        std::mem::swap(&mut self.all_postings, &mut other.all_postings);
-    }
-
-    pub(super) fn remove_posting_for_label_value(
-        &mut self,
-        label: &str,
-        value: &str,
-        ts_id: SeriesRef,
-    ) -> bool {
-        let key = IndexKey::for_label_value(label, value);
-        if let Some(bmp) = self.label_index.get_mut(&key) {
-            let removed = bmp.remove_checked(ts_id);
-            if removed && bmp.is_empty() {
-                self.label_index.remove(&key);
-            }
-            return removed;
-        }
-        false
-    }
-
-    pub(super) fn add_posting_for_label_value(
-        &mut self,
-        ts_id: SeriesRef,
-        label: &str,
-        value: &str,
-    ) -> bool {
-        self.all_postings.add(ts_id);
-        self.add_posting_for_label_value_internal(ts_id, label, value)
-    }
-
-    fn add_posting_for_label_value_internal(
-        &mut self,
-        ts_id: SeriesRef,
-        label: &str,
-        value: &str,
-    ) -> bool {
-        let key = IndexKey::for_label_value(label, value);
-        match self.label_index.entry(key) {
-            ARTEntry::Occupied(mut entry) => {
-                entry.get_mut().add(ts_id);
-                false
-            }
-            ARTEntry::Vacant(entry) => {
-                let mut bitmap = PostingsBitmap::new();
-                bitmap.add(ts_id);
-                entry.insert(bitmap);
-                true
-            }
-        }
-    }
-
-    /// Indexes a batch of loaded series in one pass. Semantically equivalent to calling
-    /// [`Postings::index_timeseries`] once per entry, but does the work set-at-a-time:
-    /// `all_postings` gets one sorted `add_many`, and the `(label=value, id)` pairs are sorted
-    /// so each posting list is touched once (`add_many` per run) with cache-friendly
-    /// sorted-order ART inserts.
-    ///
-    /// Safe against a non-empty index: existing posting lists are extended (bitmap adds are
-    /// idempotent) and `id_to_key` inserts overwrite.
-    pub(crate) fn bulk_index(&mut self, entries: Vec<BulkIndexEntry>) {
-        if entries.is_empty() {
-            return;
-        }
-
-        // (Re)indexing asserts the series are alive — revoke pending stale markings, same as
-        // `index_timeseries`.
-        if !self.stale_ids.is_empty() {
-            for entry in &entries {
-                self.stale_ids.remove(entry.id);
-            }
-        }
-
-        let mut ids: Vec<SeriesRef> = entries.iter().map(|e| e.id).collect();
-        ids.sort_unstable();
-        self.all_postings.add_many(&ids);
-
-        let total_labels: usize = entries.iter().map(|e| e.label_keys.len()).sum();
-        let mut pairs: Vec<(IndexKey, SeriesRef)> = Vec::with_capacity(total_labels);
-        for entry in entries {
-            for label_key in entry.label_keys {
-                pairs.push((label_key, entry.id));
-            }
-            self.id_to_key.insert(entry.id, entry.key);
-        }
-        pairs.sort_unstable_by(|a, b| a.0.as_bytes().cmp(b.0.as_bytes()).then(a.1.cmp(&b.1)));
-
-        // Walk the sorted pairs, flushing one add_many per run of equal label keys.
-        let mut run_ids: Vec<SeriesRef> = Vec::new();
-        let mut current_key: Option<IndexKey> = None;
-        for (key, id) in pairs {
-            match &current_key {
-                Some(cur) if *cur == key => run_ids.push(id),
-                _ => {
-                    if let Some(cur) = current_key.take() {
-                        self.add_many_for_key(cur, &run_ids);
-                    }
-                    run_ids.clear();
-                    run_ids.push(id);
-                    current_key = Some(key);
-                }
-            }
-        }
-        if let Some(cur) = current_key.take() {
-            self.add_many_for_key(cur, &run_ids);
-        }
-    }
-
-    fn add_many_for_key(&mut self, key: IndexKey, ids: &[SeriesRef]) {
-        match self.label_index.entry(key) {
-            ARTEntry::Occupied(mut entry) => entry.get_mut().add_many(ids),
-            ARTEntry::Vacant(entry) => {
-                let mut bitmap = PostingsBitmap::new();
-                bitmap.add_many(ids);
-                entry.insert(bitmap);
-            }
-        }
-    }
-
-    fn set_timeseries_key(&mut self, id: SeriesRef, new_key: &[u8]) {
-        if let Some(existing) = self.id_to_key.get(&id)
-            && existing.as_ref() == new_key
-        {
-            return;
-        }
-        let key = new_key.to_vec().into_boxed_slice();
-        self.id_to_key.insert(id, key);
-    }
-
-    pub fn index_timeseries(&mut self, ts: &TimeSeries, key: &[u8]) {
-        debug_assert!(ts.id != 0);
-        let id = ts.id;
-
-        // (Re)indexing asserts the series is alive: revoke any pending stale marking, or a later
-        // GC drain would strip the id from the very label bitmaps being filled here (e.g. the
-        // post-load repair scan re-indexing an id the reconciliation sweep just marked stale).
-        if !self.stale_ids.is_empty() {
-            self.stale_ids.remove(id);
-        }
-
-        for InternedLabel { name, value } in ts.labels.iter() {
-            self.add_posting_for_label_value_internal(id, name, value);
-        }
-
-        self.all_postings.add(id);
-        self.set_timeseries_key(id, key);
-    }
-
-    pub fn remove_timeseries(&mut self, series: &TimeSeries) -> bool {
-        let id = series.id;
-        if self.id_to_key.remove(&id).is_none() {
-            log_warning(format!(
-                "Tried to remove non-existing series id {id} from index"
-            ));
-        };
-        let removed = self.all_postings.remove_checked(id);
-        for label in series.labels.iter() {
-            self.remove_posting_for_label_value(label.name(), label.value(), id);
-        }
-        removed
-    }
-
-    pub fn count(&self) -> usize {
-        self.id_to_key.len()
-    }
-
-    pub(super) fn has_id(&self, id: SeriesRef) -> bool {
-        self.id_to_key.contains_key(&id)
-    }
-
-    /// Return postings for a key (borrowed if possible), applying stale removal.
-    /// If stale_ids is non-empty, this returns Owned.
-    fn postings_for_key(&'_ self, key: &[u8]) -> Cow<'_, PostingsBitmap> {
-        match self.label_index.get(key) {
-            Some(bmp) if self.stale_ids.is_empty() => Cow::Borrowed(bmp),
-            Some(bmp) => Cow::Owned(bmp.andnot(&self.stale_ids)),
-            None => Cow::Borrowed(&*EMPTY_BITMAP),
-        }
-    }
-
-    /// Clone postings for a key (or empty), then remove stale IDs in-place.
-    fn postings_for_key_owned(&self, key: &[u8]) -> PostingsBitmap {
-        let mut out = self.label_index.get(key).cloned().unwrap_or_default();
-        self.remove_stale_if_needed(&mut out);
-        out
-    }
-
-    pub fn postings_for_label_value<'a>(
-        &'a self,
-        name: &str,
-        value: &str,
-    ) -> Cow<'a, PostingsBitmap> {
-        let key = KeyBuffer::for_label_value(name, value);
-        self.postings_for_key(key.as_bytes())
-    }
-
-    #[inline]
-    fn remove_stale_if_needed(&self, postings: &mut PostingsBitmap) {
-        if !self.stale_ids.is_empty() {
-            postings.andnot_inplace(&self.stale_ids);
-        }
-    }
-
-    pub fn get_label_names(&self) -> BTreeSet<String> {
-        let mut names: BTreeSet<String> = BTreeSet::new();
-        for (key, map) in self.label_index.iter() {
-            if map.is_empty() {
-                continue;
-            }
-            if let Some((name, _)) = key.split()
-                && !names.contains(name)
-            {
-                names.insert(name.to_owned());
-            }
-        }
-        names
-    }
-
-    pub fn get_label_values(&self, label_name: &str) -> Vec<String> {
-        let prefix = KeyBuffer::for_prefix(label_name);
-        let mut values = Vec::with_capacity(8);
-        for (k, map) in self.label_index.prefix(prefix.as_bytes()) {
-            if !map.is_empty()
-                && let Some((_key, value)) = k.split()
-                && !value.is_empty()
-            {
-                values.push(value.to_string());
-            }
-        }
-        values
-    }
-
-    pub fn postings_for_all_label_values(&self, label_name: &str) -> PostingsBitmap {
-        let prefix = KeyBuffer::for_prefix(label_name);
-        let mut result = PostingsBitmap::new();
-        for (_, map) in self.label_index.prefix(prefix.as_bytes()) {
-            result |= map;
-        }
-        self.remove_stale_if_needed(&mut result);
-        result
-    }
-
-    /// `postings_for_label_values` returns a `PostingsBitmap` for the label pairs.
-    /// The postings here contain the ids to the series inside the index.
-    pub fn postings_for_label_values(&self, name: &str, values: &[String]) -> PostingsBitmap {
-        if let [value] = values {
-            let key = KeyBuffer::for_label_value(name, value);
-            return self.postings_for_key_owned(key.as_bytes());
-        }
-
-        let mut result = PostingsBitmap::new();
-
-        for value in values {
-            let key = KeyBuffer::for_label_value(name, value);
-            if let Some(bmp) = self.label_index.get(key.as_bytes()) {
-                result |= bmp;
-            }
-        }
-
-        self.remove_stale_if_needed(&mut result);
-        result
-    }
-
-    /// `postings_for_label_matching` returns postings having a label with the given name and a value
-    /// for which `match_fn` returns true. If no postings are found having at least one matching label,
-    /// an empty bitmap is returned.
-    pub fn postings_for_label_matching<F, STATE>(
-        &self,
-        name: &str,
-        state: &mut STATE,
-        match_fn: F,
-    ) -> PostingsBitmap
-    where
-        F: Fn(&str, &mut STATE) -> bool,
-    {
-        let prefix = KeyBuffer::for_prefix(name);
-        let start_pos = prefix.len();
-        let mut result = PostingsBitmap::new();
-
-        for (key, map) in self.label_index.prefix(prefix.as_bytes()) {
-            let value = key.sub_string(start_pos);
-            if match_fn(value, state) {
-                result |= map;
-            }
-        }
-
-        self.remove_stale_if_needed(&mut result);
-        result
-    }
-
-    /// Return all series ids corresponding to the given labels
-    pub fn postings_by_labels<T: SeriesLabel>(&self, labels: &[T]) -> PostingsBitmap {
-        let mut first = true;
-        let mut acc = PostingsBitmap::new();
-
-        for label in labels.iter() {
-            let key = KeyBuffer::for_label_value(label.name(), label.value());
-            if let Some(bmp) = self.label_index.get(key.as_bytes()) {
-                if bmp.is_empty() {
-                    break;
-                }
-                if first {
-                    acc |= bmp;
-                    first = false;
-                } else {
-                    acc &= bmp;
-                }
-            }
-        }
-        if !self.stale_ids.is_empty() {
-            acc.andnot_inplace(&self.stale_ids);
-        }
-
-        acc
-    }
-
-    /// Retrieves a `PostingsBitmap` containing postings that match the specified label and prefix.
-    ///
-    /// This function searches for postings in the internal `label_index` where the keys start with
-    /// a combination of the provided `label` and `prefix`. For each match, the function accumulates
-    /// the corresponding postings into a `PostingsBitmap`. After the accumulation, it ensures
-    /// that stale entries (if any) are removed from the resulting bitmap.
-    ///
-    /// # Parameters
-    /// - `label`: A string slice representing the label to search for in the index.
-    /// - `prefix`: A string slice representing the prefix to match against the label values in the index.
-    ///
-    /// # Returns
-    /// A `PostingsBitmap` that contains the union of all postings that match the given label and prefix.
-    pub fn postings_by_prefix(&self, label: &str, prefix: &str) -> PostingsBitmap {
-        let search_prefix = KeyBuffer::for_label_value_prefix(label, prefix);
-
-        let mut result = PostingsBitmap::new();
-        for (_key, map) in self.label_index.prefix(&search_prefix) {
-            result |= map;
-        }
-        self.remove_stale_if_needed(&mut result);
-        result
-    }
-
-    pub fn postings_by_prefix_and_predicate<F>(
-        &self,
-        label: &str,
-        prefix: &str,
-        predicate: F,
-    ) -> PostingsBitmap
-    where
-        F: Fn(&str) -> bool,
-    {
-        let search_prefix = KeyBuffer::for_label_value_prefix(label, prefix);
-
-        let mut result = PostingsBitmap::new();
-        let start_pos = label.len() + 1;
-        for (key, map) in self.label_index.prefix(&search_prefix) {
-            let value = key.sub_string(start_pos);
-            if predicate(value) {
-                result |= map;
-            }
-        }
-        self.remove_stale_if_needed(&mut result);
-
-        result
-    }
-
-    /// Get the unique series id for the given set of labels if it exists.
-    ///
-    /// This exists primarily to ensure that we disallow duplicate metric names
-    pub fn posting_id_by_labels<T: SeriesLabel>(&self, labels: &[T]) -> Option<SeriesRef> {
-        let mut it = labels.iter();
-
-        let first = it.next()?;
-        let first_key = KeyBuffer::for_label_value(first.name(), first.value());
-        let mut acc = self.label_index.get(first_key.as_bytes())?.clone();
-
-        for label in it {
-            let key = KeyBuffer::for_label_value(label.name(), label.value());
-            let bmp = self.label_index.get(key.as_bytes())?;
-            acc.and_inplace(bmp);
-            if acc.is_empty() {
-                return None;
-            }
-        }
-
-        self.remove_stale_if_needed(&mut acc);
-
-        (acc.cardinality() == 1)
-            .then(|| acc.iter().next())
-            .flatten()
-    }
-
-    pub fn postings_without_label(&'_ self, label: &str) -> Cow<'_, PostingsBitmap> {
-        let to_remove = self.postings_for_all_label_values(label);
-        if to_remove.is_empty() {
-            Cow::Borrowed(&self.all_postings)
-        } else {
-            Cow::Owned(self.all_postings.andnot(&to_remove))
-        }
-    }
-
-    fn postings_for_any_of_labels(&self, labels: &[&str]) -> PostingsBitmap {
-        let mut to_remove = PostingsBitmap::new();
-        for label in labels {
-            let prefix = KeyBuffer::for_prefix(label);
-            for (_, map) in self.label_index.prefix(prefix.as_bytes()) {
-                to_remove.or_inplace(map);
-            }
-        }
-        self.remove_stale_if_needed(&mut to_remove);
-        to_remove
-    }
-
-    pub fn postings_for_filter(
-        &'_ self,
-        filter: &LabelFilter,
-    ) -> ValkeyResult<Cow<'_, PostingsBitmap>> {
-        let result = match filter.matcher {
-            PredicateMatch::Equal(ref value) => handle_equal_match(self, &filter.label, value),
-            PredicateMatch::NotEqual(ref value) => {
-                handle_not_equal_match(self, &filter.label, value)
-            }
-            PredicateMatch::MatchAll => Cow::Borrowed(&self.all_postings),
-            PredicateMatch::MatchNone => Cow::Borrowed(&*EMPTY_BITMAP),
-            PredicateMatch::RegexEqual(_) => handle_regex_equal_match(self, filter)?,
-            PredicateMatch::RegexNotEqual(_) => handle_regex_not_equal_match(self, filter)?,
-            PredicateMatch::StartsWith(ref prefix) => {
-                handle_starts_with(self, &filter.label, prefix)
-            }
-            PredicateMatch::NotStartsWith(ref prefix) => {
-                handle_not_starts_with(self, &filter.label, prefix)
-            }
-            PredicateMatch::Contains(ref needle) => handle_contains(self, &filter.label, needle),
-            PredicateMatch::NotContains(_) => handle_not_contains(self, filter)?,
-        };
-        Ok(result)
-    }
-
-    fn inverse_postings_for_filter(&'_ self, filter: &LabelFilter) -> Cow<'_, PostingsBitmap> {
-        match &filter.matcher {
-            PredicateMatch::NotEqual(pv) => handle_equal_match(self, &filter.label, pv),
-            // If the matcher being inverted is ="", we just want all the values.
-            PredicateMatch::Equal(PredicateValue::String(s)) if s.is_empty() => {
-                Cow::Owned(self.postings_for_all_label_values(&filter.label))
-            }
-            PredicateMatch::MatchAll => Cow::Borrowed(&*EMPTY_BITMAP),
-            PredicateMatch::MatchNone => Cow::Borrowed(&self.all_postings),
-            // If the matcher being inverted is =~"", we just want all the values.
-            PredicateMatch::RegexEqual(re) if matches!(re.regex.as_str(), "" | ".*") => {
-                Cow::Owned(self.postings_for_all_label_values(&filter.label))
-            }
-            PredicateMatch::StartsWith(prefix) => {
-                let all = self.postings_for_all_label_values(&filter.label);
-                let matching_prefix = postings_by_prefix_value(self, &filter.label, prefix);
-                Cow::Owned(all.andnot(&matching_prefix))
-            }
-            PredicateMatch::NotStartsWith(prefix) => {
-                Cow::Owned(postings_by_prefix_value(self, &filter.label, prefix))
-            }
-            PredicateMatch::Contains(needle) => {
-                let all = self.postings_for_all_label_values(&filter.label);
-                let matching = handle_contains(self, &filter.label, needle).into_owned();
-                Cow::Owned(all.andnot(&matching))
-            }
-            PredicateMatch::NotContains(needle) => {
-                Cow::Owned(handle_contains(self, &filter.label, needle).into_owned())
-            }
-            _ => {
-                let mut state = filter;
-                let postings =
-                    self.postings_for_label_matching(&filter.label, &mut state, |s, state| {
-                        let valid = state.matches(s);
-                        !valid
-                    });
-                Cow::Owned(postings)
-            }
-        }
-    }
-
-    /// `postings_for_label_filters` assembles a single postings iterator against the index
-    /// based on the given matchers.
-    pub fn postings_for_label_filters(
-        &'_ self,
-        filters: &[LabelFilter],
-    ) -> ValkeyResult<Cow<'_, PostingsBitmap>> {
-        if filters.is_empty() {
-            return Ok(Cow::Borrowed(&self.all_postings));
-        }
-        if filters.len() == 1 {
-            let filter = &filters[0];
-            // follow Prometheus here: if we have an empty matcher and label, return all postings.
-            if filter.label.is_empty() && filter.matcher.is_empty() {
-                return Ok(Cow::Borrowed(&self.all_postings));
-            }
-            // shortcut the handling of simple equality matchers
-            if !filter.is_negative_matcher() && !filter.matches_empty() {
-                let it = self.postings_for_filter(filter)?;
-                if it.is_empty() {
-                    return Ok(Cow::Borrowed(&*EMPTY_BITMAP));
-                }
-                return Ok(it);
-            }
-        }
-
-        let mut its: SmallVec<[_; 4]> = SmallVec::new();
-        let mut not_its: SmallVec<[Cow<PostingsBitmap>; 4]> = SmallVec::new();
-
-        let mut sorted_matchers: SmallVec<[(&LabelFilter, bool, bool); 4]> = SmallVec::new();
-
-        let mut has_subtracting_matchers = false;
-        let mut has_intersecting_matchers = false;
-        for m in filters {
-            let matches_empty = m.matches("");
-
-            let is_subtracting = matches_empty || m.is_negative_matcher();
-
-            if is_subtracting {
-                has_subtracting_matchers = true;
-            } else {
-                has_intersecting_matchers = true;
-            }
-
-            sorted_matchers.push((m, matches_empty, is_subtracting))
-        }
-
-        if has_subtracting_matchers && !has_intersecting_matchers {
-            // If there's nothing to subtract from, add in everything and remove the not_its later.
-            // We prefer to get all_postings so that the base of subtraction (i.e., all_postings)
-            // doesn't include series that may be added to the index reader during this function call.
-            its.push(Cow::Borrowed(&self.all_postings));
-        };
-
-        // Sort matchers to have the intersecting matchers first.
-        // This way the base for subtraction is smaller, and there is no chance that the set we subtract
-        // from contains postings of series that didn't exist when we constructed the set we subtract by.
-        sorted_matchers.sort_by(|i, j| -> Ordering {
-            let is_i_subtracting = i.2;
-            let is_j_subtracting = j.2;
-            if !is_i_subtracting && is_j_subtracting {
-                return Ordering::Less;
-            }
-            // sort by match cost
-            let cost_i = i.0.cost();
-            let cost_j = j.0.cost();
-            cost_i.cmp(&cost_j)
-        });
-
-        for (filter, matches_empty, _is_subtracting) in sorted_matchers {
-            //let value = &m.value;
-            let name = &filter.label;
-
-            if name.is_empty() && matches_empty {
-                // We already handled the case at the top of the function,
-                // and it is unexpected to get all postings again here.
-                return Err(ValkeyError::Str(MISSING_FILTER));
-            }
-
-            if matches!(filter.matcher, PredicateMatch::MatchAll) {
-                continue;
-            }
-
-            if matches!(filter.matcher, PredicateMatch::MatchNone) {
-                return Ok(Cow::Borrowed(&*EMPTY_BITMAP));
-            }
-
-            let typ = filter.op();
-            let regex_value = filter.regex_text().unwrap_or("");
-
-            match (typ, regex_value) {
-                // .* regexp matches any string: do nothing
-                (MatchOp::RegexEqual, ".*") => continue,
-
-                // .* regexp does not match any string: return empty
-                (MatchOp::RegexNotEqual, ".*") => {
-                    return Ok(Cow::Borrowed(&*EMPTY_BITMAP));
-                }
-
-                // .+ regexp matches any non-empty string
-                (MatchOp::RegexEqual, ".+") => {
-                    // .+ regexp matches any non-empty string: get postings for all label values.
-                    let it = self.postings_for_all_label_values(&filter.label);
-                    its.push(Cow::Owned(it));
-                }
-
-                // .+ regexp does not match any non-empty string
-                (MatchOp::RegexNotEqual, ".+") => {
-                    let it = self.postings_for_all_label_values(&filter.label);
-                    not_its.push(Cow::Owned(it));
-                }
-                // See which label must be non-empty.
-                // Optimization for a case like {l=~".", l!="1"}.
-                _ if !matches_empty => {
-                    // If this matcher must be non-empty, we can be smarter.
-                    let is_not = matches!(
-                        typ,
-                        MatchOp::NotEqual
-                            | MatchOp::RegexNotEqual
-                            | MatchOp::NotContains
-                            | MatchOp::NotStartsWith
-                    );
-                    match (is_not, matches_empty) {
-                        // l!="foo"
-                        (true, true) => {
-                            // If the label can't be empty and is a Not and the inner matcher
-                            // doesn't match empty, then subtract it out at the end.
-                            let inverse = filter.clone().inverse();
-                            let it = self.postings_for_filter(&inverse)?;
-                            not_its.push(it);
-                        }
-                        // l!=""
-                        (true, false) => {
-                            // If the label can't be empty and is a Not, but the inner matcher can
-                            // be empty, we need to use inverse_postings_for_filter.
-                            let inverse = filter.clone().inverse();
-                            let it = self.inverse_postings_for_filter(&inverse);
-                            if it.is_empty() {
-                                return Ok(Cow::Borrowed(&*EMPTY_BITMAP));
-                            }
-                            its.push(it);
-                        }
-                        // l="a", l=~"a|b", etc.
-                        _ => {
-                            // Non-Not matcher, use normal postings_for_filter.
-                            let it = self.postings_for_filter(filter)?;
-                            if it.is_empty() {
-                                return Ok(Cow::Borrowed(&*EMPTY_BITMAP));
-                            }
-                            its.push(it);
-                        }
-                    }
-                }
-                _ => {
-                    // l=""
-                    // If the matchers for a label name selects an empty value, it selects all
-                    // the series which also don't have the label name. See:
-                    // https://github.com/prometheus/prometheus/issues/3575 and
-                    // https://github.com/prometheus/prometheus/pull/3578#issuecomment-351653555
-                    let it = self.inverse_postings_for_filter(filter);
-
-                    not_its.push(it)
-                }
-            }
-        }
-
-        // optimization: if we have a single iterator and no not_its, return it directly, saving a clone.
-        if its.len() == 1 && not_its.is_empty() {
-            let single = its
-                .pop()
-                .expect("unexpected out of bounds error running matchers");
-
-            return Ok(single);
-        }
-
-        let mut result = if its.is_empty() {
-            self.all_postings.clone()
-        } else {
-            // sort by cardinality first to reduce the amount of work
-            its.sort_by_key(|a| a.cardinality());
-            intersection(its)
-        };
-
-        for not in not_its {
-            result.andnot_inplace(&not)
-        }
-
-        Ok(Cow::Owned(result))
-    }
-
-    pub fn postings_for_selector(
-        &'_ self,
-        selector: &SeriesSelector,
-    ) -> ValkeyResult<Cow<'_, PostingsBitmap>> {
-        match &selector {
-            SeriesSelector::And(filters) => self.postings_for_label_filters(filters),
-            SeriesSelector::Or(filters) => self.process_or_matchers(filters),
-        }
-    }
-
-    pub fn postings_for_selectors(
-        &'_ self,
-        selectors: &[SeriesSelector],
-    ) -> ValkeyResult<Cow<'_, PostingsBitmap>> {
-        match selectors {
-            [] => Ok(Cow::Borrowed(&*EMPTY_BITMAP)),
-            [selector] => {
-                let result = self.postings_for_selector(selector)?;
-                if !self.stale_ids.is_empty() {
-                    let mut result = result.into_owned();
-                    result.andnot_inplace(&self.stale_ids);
-                    return Ok(Cow::Owned(result));
-                }
-                Ok(result)
-            }
-            _ => {
-                let first = self.postings_for_selector(&selectors[0])?;
-
-                let mut result = first.into_owned();
-                for selector in &selectors[1..] {
-                    let bitmap = self.postings_for_selector(selector)?;
-                    result.and_inplace(&bitmap);
-                }
-
-                if !self.stale_ids.is_empty() {
-                    result.andnot_inplace(&self.stale_ids);
-                }
-
-                Ok(Cow::Owned(result))
-            }
-        }
-    }
-
-    fn process_or_matchers(
-        &'_ self,
-        filters: &[FilterList],
-    ) -> ValkeyResult<Cow<'_, PostingsBitmap>> {
-        match filters {
-            [] => Ok(Cow::Borrowed(&self.all_postings)),
-            [filters] => self.postings_for_label_filters(filters),
-            _ => {
-                let mut result = PostingsBitmap::new();
-                // maybe chili here to run in parallel
-                for matchers in filters {
-                    let postings = self.postings_for_label_filters(matchers)?;
-                    result.or_inplace(&postings);
-                }
-                Ok(Cow::Owned(result))
-            }
-        }
-    }
-
-    pub(crate) fn get_key_by_id(&self, id: SeriesRef) -> Option<&KeyType> {
-        self.id_to_key.get(&id)
-    }
-
-    pub(crate) fn mark_id_as_stale(&mut self, id: SeriesRef) {
-        self.mark_ids_as_stale(&[id]);
-    }
-
-    /// Marks ids as stale by adding its ID to the stale IDs set.
-    /// ## Context
-    /// This is used in the case of possible index sync issues. When the index is queried and an id is returned
-    /// with no corresponding series, we have no access to the series data to do a proper cleanup. Also, at the end
-    /// of cluster migration, the source node needs to remove all series ids corresponding to the migrated slots.
-    ///
-    /// We remove the key from the index and mark the ID as stale, which will be cleaned up later in a background task.
-    /// The stale IDs are stored in a bitmap for efficient removal and are checked to ensure that no stale IDs are
-    /// returned in queries until they are removed.
-    pub(crate) fn mark_ids_as_stale(&mut self, ids: &[SeriesRef]) {
-        for id in ids {
-            let _ = self.id_to_key.remove(id);
-        }
-        self.stale_ids.add_many(ids);
-        self.all_postings.remove_many(ids);
-    }
-
-    #[cfg(test)]
-    pub(super) fn has_stale_ids(&self) -> bool {
-        !self.stale_ids.is_empty()
-    }
-
-    /// Removes stale series IDs from a subset of the index structures.
-    ///
-    /// This method processes at most `count` keys starting from `start_prefix`,
-    /// removing stale IDs from their bitmaps and cleaning up empty entries.
-    ///
-    /// ## Arguments
-    /// * `start_prefix` - The key to start processing from (inclusive)
-    /// * `count` - Maximum number of keys to process in this batch
-    ///
-    /// ## Returns
-    /// * `Option<IndexKey>` - The next key to continue processing from, or None if processing is complete
-    ///
-    pub(crate) fn remove_stale_ids(
-        &mut self,
-        start_prefix: Option<IndexKey>,
-        count: usize,
-    ) -> Option<IndexKey> {
-        // Skip if there are no stale IDs to process
-        if self.stale_ids.is_empty() {
-            return None;
-        }
-
-        let mut keys_processed = 0;
-        let mut keys_to_remove = Vec::new();
-        let mut next_key = None;
-
-        // Resume from the cursor key using an ordered range (NOT a prefix scan): a full key used
-        // as a prefix only matches itself and its extensions, which would abandon the remaining
-        // keys and prematurely clear `stale_ids`.
-        let range = match start_prefix {
-            Some(k) => (Bound::Included(k), Bound::Unbounded),
-            None => (Bound::Unbounded, Bound::Unbounded),
-        };
-
-        for (key, bitmap) in self.label_index.range_mut::<IndexKey, _>(range) {
-            // Remove stale IDs from the bitmap
-            let should_remove = if !bitmap.is_empty() {
-                bitmap.andnot_inplace(&self.stale_ids);
-                bitmap.is_empty()
-            } else {
-                true
-            };
-
-            if should_remove {
-                keys_to_remove.push(key.clone());
-            }
-
-            if keys_processed == count {
-                // Save the key we stopped at as the next starting point
-                next_key = Some(key.clone());
-                break;
-            }
-
-            keys_processed += 1;
-        }
-
-        // Process empty keys
-        for key in keys_to_remove {
-            self.label_index.remove(&key);
-        }
-
-        // If we processed some keys but didn't reach the end, we can return the next key to continue from.
-        // If we processed keys and there are no more keys to process, it means we've reached the end of the index,
-        // so we can clear the stale IDs as they have been fully processed.
-        if keys_processed > 0 && next_key.is_none() {
-            self.stale_ids.clear();
-        }
-
-        next_key
-    }
-
-    /// Incrementally optimizes posting bitmaps for better memory usage and performance.
-    ///
-    /// This method processes at most `count` keys starting from `start_prefix`,
-    /// performing the following optimizations on each bitmap:
-    /// 1. Remove the bitmap if it is empty
-    /// 2. Call run_optimize() to optimize the bitmap's internal structure
-    /// 3. Call shrink_to_fit() to reduce memory overhead
-    ///
-    /// ### Arguments
-    /// * `start_prefix` - The key to start processing from (inclusive)
-    /// * `count` - Maximum number of keys to process in this batch
-    ///
-    /// ### Returns
-    /// * `Option<IndexKey>` - The next key to continue processing from, or None if processing is complete
-    ///
-    pub(crate) fn optimize_postings(
-        &mut self,
-        start_prefix: Option<IndexKey>,
-        count: usize,
-    ) -> Option<IndexKey> {
-        let mut next_key = None;
-
-        if start_prefix.is_none() {
-            optimize_bitmap(&mut self.all_postings);
-        }
-
-        let mut keys_to_delete = Vec::new();
-        let mut keys_processed: usize = 0;
-        // Resume from the cursor key using an ordered range rather than a prefix scan (a full key
-        // as a prefix only matches itself and its extensions).
-        let range = match start_prefix {
-            Some(k) => (Bound::Included(k), Bound::Unbounded),
-            None => (Bound::Unbounded, Bound::Unbounded),
-        };
-
-        // Collect keys to process
-        for (key, bitmap) in self.label_index.range_mut::<IndexKey, _>(range) {
-            if bitmap.is_empty() {
-                keys_to_delete.push(key.clone());
-                continue;
-            }
-            if keys_processed == count {
-                // Save the key we stopped at as the next starting point
-                next_key = Some(key.clone());
-                break;
-            }
-
-            optimize_bitmap(bitmap);
-
-            keys_processed += 1;
-        }
-
-        // Remove empty bitmaps collected earlier
-        for key in keys_to_delete {
-            self.label_index.remove(&key);
-        }
-
-        next_key
-    }
-}
-
-/// Optimizes a bitmap in place for better memory usage and performance.
-/// This applies run_optimize() and shrink_to_fit() operations to the bitmap
-/// if it exists in the index.
-fn optimize_bitmap(bitmap: &mut PostingsBitmap) {
-    // Optimize the bitmap's internal structure
-    bitmap.run_optimize();
-
-    // Shrink to fit to reduce memory overhead
-    bitmap.shrink_to_fit();
-}
-
-fn handle_equal_match<'a>(
-    ix: &'a Postings,
-    label: &str,
-    value: &PredicateValue,
-) -> Cow<'a, PostingsBitmap> {
-    match value {
-        PredicateValue::String(s) => {
-            if s.is_empty() {
-                return ix.postings_without_label(label);
-            }
-            ix.postings_for_label_value(label, s)
-        }
-        PredicateValue::List(val) => match val.len() {
-            0 => ix.postings_without_label(label),
-            1 => {
-                if val[0].is_empty() {
-                    ix.postings_without_label(label)
-                } else {
-                    ix.postings_for_label_value(label, &val[0])
-                }
-            }
-            _ => {
-                // If the list contains an explicit empty alternative, include series
-                // without the label as well.
-                let contains_empty = val.iter().any(|s| s.is_empty());
-
-                let non_empty_values: Vec<String> =
-                    val.iter().filter(|s| !s.is_empty()).cloned().collect();
-
-                if non_empty_values.is_empty() {
-                    // only empty alternative -> postings without label
-                    return ix.postings_without_label(label);
-                }
-
-                let mut result = ix.postings_for_label_values(label, &non_empty_values);
-                // include series that don't have the label only if the original
-                // alternatives contained an empty branch.
-                if contains_empty {
-                    let without = ix.postings_without_label(label).into_owned();
-                    result |= without;
-                }
-                Cow::Owned(result)
-            }
-        },
-        PredicateValue::Empty => ix.postings_without_label(label),
-    }
-}
-
-// return postings for series which has the label `label
-fn with_label<'a>(ix: &'a Postings, label: &str) -> Cow<'a, PostingsBitmap> {
-    let mut state = ();
-    let postings = ix.postings_for_label_matching(label, &mut state, |_value, _| true);
-    Cow::Owned(postings)
-}
-
-fn handle_not_equal_match<'a>(
-    ix: &'a Postings,
-    label: &str,
-    value: &PredicateValue,
-) -> Cow<'a, PostingsBitmap> {
-    // the time series has a label named label
-    match value {
-        PredicateValue::String(s) => {
-            if s.is_empty() {
-                return with_label(ix, label);
-            }
-            let all = &ix.all_postings;
-            let postings = ix.postings_for_label_value(label, s);
-            if postings.is_empty() {
-                Cow::Borrowed(all)
-            } else {
-                let result = all.andnot(&postings);
-                Cow::Owned(result)
-            }
-        }
-        PredicateValue::List(values) => {
-            match values.len() {
-                0 => with_label(ix, label),
-                _ => {
-                    // get postings for label m.label without values in values
-                    let to_remove = ix.postings_for_label_values(label, values);
-                    let all_postings = &ix.all_postings;
-                    if to_remove.is_empty() {
-                        Cow::Borrowed(all_postings)
-                    } else {
-                        let result = all_postings.andnot(&to_remove);
-                        Cow::Owned(result)
-                    }
-                }
-            }
-        }
-        PredicateValue::Empty => with_label(ix, label),
-    }
-}
-
-#[inline]
-fn postings_matching_filter(postings: &Postings, filter: &LabelFilter) -> PostingsBitmap {
-    let mut state = filter;
-    postings.postings_for_label_matching(&filter.label, &mut state, |value, f| f.matches(value))
-}
-
-fn handle_regex_equal_match<'a>(
-    postings: &'a Postings,
-    filter: &LabelFilter,
-) -> ValkeyResult<Cow<'a, PostingsBitmap>> {
-    if filter.matches_empty() {
-        return Ok(postings.postings_without_label(&filter.label));
-    }
-    // The caller dispatches on this same enum (`postings_for_filter`), so this arm is only
-    // reached for `RegexEqual` filters; a mismatch means that invariant broke.
-    let PredicateMatch::RegexEqual(re) = &filter.matcher else {
-        debug_assert!(false, "unexpected matcher type in handle_regex_equal_match");
-        return Err(ValkeyError::Str(INTERNAL_ERROR));
-    };
-    let res = if let Some(prefix) = &re.prefix {
-        postings.postings_by_prefix_and_predicate(&filter.label, prefix, |v| re.is_match(v))
-    } else {
-        let mut state = ();
-        postings
-            .postings_for_label_matching(&filter.label, &mut state, |value, _| re.is_match(value))
-    };
-    Ok(Cow::Owned(res))
-}
-
-fn handle_regex_not_equal_match<'a>(
-    postings: &'a Postings,
-    filter: &LabelFilter,
-) -> ValkeyResult<Cow<'a, PostingsBitmap>> {
-    if filter.matches_empty() {
-        return Ok(with_label(postings, &filter.label));
-    }
-    // The caller dispatches on this same enum (`postings_for_filter`), so this arm is only
-    // reached for `RegexNotEqual` filters; a mismatch means that invariant broke.
-    let PredicateMatch::RegexNotEqual(re) = &filter.matcher else {
-        debug_assert!(
-            false,
-            "unexpected matcher type in handle_regex_not_equal_match"
-        );
-        return Err(ValkeyError::Str(INTERNAL_ERROR));
-    };
-    let mut state = ();
-    let res = postings
-        .postings_for_label_matching(&filter.label, &mut state, |value, _| !re.is_match(value));
-    Ok(Cow::Owned(res))
-}
-
-fn handle_starts_with<'a>(
-    postings: &'a Postings,
-    label: &str,
-    prefix: &PredicateValue,
-) -> Cow<'a, PostingsBitmap> {
-    Cow::Owned(postings_by_prefix_value(postings, label, prefix))
-}
-
-fn handle_not_starts_with<'a>(
-    postings: &'a Postings,
-    label: &str,
-    prefix: &PredicateValue,
-) -> Cow<'a, PostingsBitmap> {
-    let matching_prefix = postings_by_prefix_value(postings, label, prefix);
-    Cow::Owned(postings.all_postings.andnot(&matching_prefix))
-}
-
-fn postings_by_prefix_value(
-    postings: &Postings,
-    label: &str,
-    value: &PredicateValue,
-) -> PostingsBitmap {
-    match value {
-        PredicateValue::Empty => PostingsBitmap::new(),
-        PredicateValue::String(prefix) => postings.postings_by_prefix(label, prefix),
-        PredicateValue::List(prefixes) => {
-            let mut result = PostingsBitmap::new();
-            for prefix in prefixes {
-                result.or_inplace(&postings.postings_by_prefix(label, prefix));
-            }
-            result
-        }
-    }
-}
-
-fn handle_contains<'a>(
-    postings: &'a Postings,
-    label: &str,
-    value: &PredicateValue,
-) -> Cow<'a, PostingsBitmap> {
-    let mut state = value;
-    let res = postings.postings_for_label_matching(label, &mut state, |candidate, value| {
-        value.matches_contains_any(candidate)
-    });
-    Cow::Owned(res)
-}
-
-fn handle_not_contains<'a>(
-    postings: &'a Postings,
-    filter: &LabelFilter,
-) -> ValkeyResult<Cow<'a, PostingsBitmap>> {
-    // The caller dispatches on this same enum (`postings_for_filter`), so this arm is only
-    // reached for `NotContains` filters; a mismatch means that invariant broke.
-    let PredicateMatch::NotContains(value) = &filter.matcher else {
-        debug_assert!(false, "unexpected matcher type in handle_not_contains");
-        return Err(ValkeyError::Str(INTERNAL_ERROR));
-    };
-    let mut state = value;
-    let contains_matches =
-        postings.postings_for_label_matching(&filter.label, &mut state, |candidate, value| {
-            value.matches_contains_any(candidate)
-        });
-    let res = &postings.all_postings;
-    let value = res.andnot(&contains_matches);
-    Ok(Cow::Owned(value))
-}
-
-fn intersection<'a, I>(its: I) -> PostingsBitmap
-where
-    I: IntoIterator<Item = Cow<'a, PostingsBitmap>>,
-{
-    let mut its = its.into_iter();
-    if let Some(it) = its.next() {
-        let mut result = it.into_owned();
-
-        for it in its {
-            if it.is_empty() {
-                result.clear();
-                return result;
-            }
-
-            result.and_inplace(&it);
-        }
-
-        result
-    } else {
-        PostingsBitmap::new()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::labels::{Label, MetricName};
-    use crate::series::time_series::TimeSeries;
-
-    fn make_series(id: SeriesRef, labels: &[(&str, &str)]) -> TimeSeries {
-        let mut series = TimeSeries::new();
-        series.id = id;
-        let labels: Vec<Label> = labels
-            .iter()
-            .map(|(name, value)| Label::new(*name, *value))
-            .collect();
-        series.labels = MetricName::new(&labels);
-        series
-    }
-
-    fn make_bulk_entry(id: SeriesRef, key: &str, labels: &[(&str, &str)]) -> BulkIndexEntry {
-        BulkIndexEntry {
-            id,
-            key: key.as_bytes().to_vec().into_boxed_slice(),
-            label_keys: labels
-                .iter()
-                .map(|(name, value)| IndexKey::for_label_value(name, value))
-                .collect(),
-        }
-    }
-
-    /// Asserts the two indexes are structurally identical (same label keys, same posting-list
-    /// contents, same id map, same all_postings).
-    fn assert_postings_eq(a: &Postings, b: &Postings) {
-        assert_eq!(a.id_to_key, b.id_to_key);
-        assert_eq!(a.all_postings, b.all_postings, "all_postings differ");
-        let a_entries: Vec<_> = a.label_index.iter().collect();
-        let b_entries: Vec<_> = b.label_index.iter().collect();
-        assert_eq!(a_entries.len(), b_entries.len(), "label key counts differ");
-        for ((ka, va), (kb, vb)) in a_entries.iter().zip(b_entries.iter()) {
-            assert_eq!(ka.as_str(), kb.as_str());
-            assert_eq!(va, vb, "posting list for {} differs", ka.as_str());
-        }
-    }
-
-    type SeriesSpec = (SeriesRef, String, Vec<(String, String)>);
-
-    #[test]
-    fn test_bulk_index_matches_per_key_indexing() {
-        // 40 series over overlapping labels: shared job/env values, unique instance values.
-        let specs: Vec<SeriesSpec> = (1..=40u64)
-            .map(|i| {
-                let key = format!("ts:{i}");
-                let labels = vec![
-                    ("job".to_string(), format!("job{}", i % 3)),
-                    (
-                        "env".to_string(),
-                        if i % 2 == 0 { "prod" } else { "dev" }.to_string(),
-                    ),
-                    ("instance".to_string(), format!("host-{i}")),
-                ];
-                (i, key, labels)
-            })
-            .collect();
-
-        let mut per_key = Postings::default();
-        for (id, key, labels) in &specs {
-            let labels: Vec<(&str, &str)> = labels
-                .iter()
-                .map(|(n, v)| (n.as_str(), v.as_str()))
-                .collect();
-            let series = make_series(*id, &labels);
-            per_key.index_timeseries(&series, key.as_bytes());
-        }
-
-        let mut bulk = Postings::default();
-        let entries: Vec<BulkIndexEntry> = specs
-            .iter()
-            .map(|(id, key, labels)| {
-                let labels: Vec<(&str, &str)> = labels
-                    .iter()
-                    .map(|(n, v)| (n.as_str(), v.as_str()))
-                    .collect();
-                make_bulk_entry(*id, key, &labels)
-            })
-            .collect();
-        bulk.bulk_index(entries);
-
-        assert_postings_eq(&per_key, &bulk);
-    }
-
-    #[test]
-    fn test_bulk_index_merges_into_non_empty_index() {
-        // Pre-populate per-key (as in a degraded window: some keys per-key, some bulk).
-        let mut postings = Postings::default();
-        let series = make_series(1, &[("job", "web"), ("host", "a")]);
-        postings.index_timeseries(&series, b"key1");
-
-        postings.bulk_index(vec![
-            make_bulk_entry(2, "key2", &[("job", "web"), ("host", "b")]),
-            make_bulk_entry(3, "key3", &[("job", "batch"), ("host", "a")]),
-        ]);
-
-        let web = postings.postings_for_label_value("job", "web");
-        assert!(web.contains(1) && web.contains(2) && !web.contains(3));
-        let host_a = postings.postings_for_label_value("host", "a");
-        assert!(host_a.contains(1) && host_a.contains(3));
-        assert_eq!(postings.count(), 3);
-        assert!(postings.all_postings.contains(2));
-    }
-
-    #[test]
-    fn test_bulk_index_revokes_stale_marks() {
-        let mut postings = Postings::default();
-        let series = make_series(7, &[("job", "web")]);
-        postings.index_timeseries(&series, b"key7");
-        postings.mark_id_as_stale(7);
-        assert!(postings.has_stale_ids());
-
-        // Re-loading the id asserts it is alive again (same semantics as index_timeseries).
-        postings.bulk_index(vec![make_bulk_entry(7, "key7", &[("job", "web")])]);
-
-        assert!(!postings.stale_ids.contains(7));
-        assert!(postings.all_postings.contains(7));
-        assert!(postings.postings_for_label_value("job", "web").contains(7));
-    }
-
-    #[test]
-    fn test_memory_postings_add_and_remove() {
-        let mut postings = Postings::default();
-
-        // Add postings
-        postings.add_posting_for_label_value(1, "label1", "value1");
-        postings.add_posting_for_label_value(1, "label2", "value2");
-        postings.add_posting_for_label_value(2, "label1", "value1");
-
-        // Check postings
-        assert_eq!(
-            postings
-                .postings_for_label_value("label1", "value1")
-                .cardinality(),
-            2
-        );
-        assert_eq!(
-            postings
-                .postings_for_label_value("label2", "value2")
-                .cardinality(),
-            1
-        );
-
-        // Remove posting
-        postings.remove_posting_for_label_value("label1", "value1", 1);
-        assert_eq!(
-            postings
-                .postings_for_label_value("label1", "value1")
-                .cardinality(),
-            1
-        );
-
-        // Remove non-existent posting (should not panic)
-        postings.remove_posting_for_label_value("label3", "value3", 3);
-    }
-
-    #[test]
-    fn test_postings_multiple_values_same_label() {
-        let mut postings = Postings::default();
-
-        // Add postings for multiple values of the same label
-        postings.add_posting_for_label_value(1, "label1", "value1");
-        postings.add_posting_for_label_value(2, "label1", "value2");
-        postings.add_posting_for_label_value(3, "label1", "value3");
-        postings.add_posting_for_label_value(4, "label1", "value1");
-
-        // Query for multiple values of the same label
-        let values = vec!["value1".to_string(), "value3".to_string()];
-        let result = postings.postings_for_label_values("label1", &values);
-
-        // Check that the result contains the correct series IDs
-        assert_eq!(result.cardinality(), 3);
-        assert!(result.contains(1));
-        assert!(result.contains(3));
-        assert!(result.contains(4));
-        assert!(!result.contains(2));
-    }
-
-    #[test]
-    fn test_postings_with_duplicate_values() {
-        let mut postings = Postings::default();
-
-        // Add some postings
-        postings.add_posting_for_label_value(1, "label", "value1");
-        postings.add_posting_for_label_value(2, "label", "value2");
-        postings.add_posting_for_label_value(3, "label", "value1");
-
-        // Create an array with duplicate values
-        let values = vec![
-            "value1".to_string(),
-            "value2".to_string(),
-            "value1".to_string(),
-        ];
-
-        // Call the postings method
-        let result = postings.postings_for_label_values("label", &values);
-
-        // Check the result
-        assert_eq!(result.cardinality(), 3);
-        assert!(result.contains(1));
-        assert!(result.contains(2));
-        assert!(result.contains(3));
-    }
-
-    #[test]
-    fn test_equal_list_includes_empty_alternative() {
-        // Ensure that an equality list containing an explicit empty alternative
-        // includes series that do not have the label (postings_without_label).
-        let mut postings = Postings::default();
-
-        // Series 1 has label "i" with value "x"
-        postings.add_posting_for_label_value(1, "i", "x");
-
-        // Series 2 exists in all_postings but does not have label "i"
-        postings.all_postings.add(2);
-
-        // Create a LabelFilter for i in ("x", "")
-        let lf = LabelFilter {
-            label: "i".to_string(),
-            matcher: PredicateMatch::Equal(PredicateValue::from(vec![
-                "x".to_string(),
-                "".to_string(),
-            ])),
-        };
-
-        let res = postings.postings_for_label_filters(&[lf]).unwrap();
-        let res = res.into_owned();
-
-        // Both series 1 (has value x) and 2 (no label) should be present.
-        assert!(res.contains(1));
-        assert!(res.contains(2));
-    }
-
-    #[test]
-    fn test_postings_all_values_match() {
-        let mut postings = Postings::default();
-
-        // Add some postings
-        postings.add_posting_for_label_value(1, "label", "value1");
-        postings.add_posting_for_label_value(2, "label", "value2");
-        postings.add_posting_for_label_value(3, "label", "value3");
-        postings.add_posting_for_label_value(4, "label", "value1");
-
-        // Create values to search for
-        let values = vec![
-            "value1".to_string(),
-            "value2".to_string(),
-            "value3".to_string(),
-        ];
-
-        // Get the postings
-        let result = postings.postings_for_label_values("label", &values);
-
-        // Check if the result contains all the expected series IDs
-        assert_eq!(result.cardinality(), 4);
-        assert!(result.contains(1));
-        assert!(result.contains(2));
-        assert!(result.contains(3));
-        assert!(result.contains(4));
-    }
-
-    #[test]
-    fn test_postings_with_large_number_of_values() {
-        let mut postings = Postings::default();
-        let label_name = "large_label";
-        let num_values = 10_000;
-
-        // Add postings for a large number of values
-        for i in 0..num_values {
-            postings.add_posting_for_label_value(i as SeriesRef, label_name, &format!("value_{i}"));
-        }
-
-        // Create a large array of values to search for
-        let values: Vec<String> = (0..num_values).map(|i| format!("value_{i}")).collect();
-
-        // Measure the time taken to execute the postings function
-        let start_time = std::time::Instant::now();
-        let result = postings.postings_for_label_values(label_name, &values);
-        let duration = start_time.elapsed();
-
-        // Assert that all series IDs are present in the result
-        assert_eq!(result.cardinality() as usize, num_values);
-        for i in 0..num_values {
-            assert!(result.contains(i as SeriesRef));
-        }
-
-        // Check that the execution time is reasonable (adjust the threshold as needed)
-        assert!(
-            duration < std::time::Duration::from_secs(1),
-            "Postings retrieval took too long: {duration:?}"
-        );
-    }
-
-    #[test]
-    fn test_postings_with_unicode_characters() {
-        let mut postings = Postings::default();
-
-        // Add postings with Unicode characters
-        postings.add_posting_for_label_value(1, "标签", "值1");
-        postings.add_posting_for_label_value(2, "标签", "値2");
-        postings.add_posting_for_label_value(3, "标签", "🌟");
-
-        // Test postings method with Unicode characters
-        let values = vec!["值1".to_string(), "値2".to_string(), "🌟".to_string()];
-        let result = postings.postings_for_label_values("标签", &values);
-
-        assert_eq!(result.cardinality(), 3);
-        assert!(result.contains(1));
-        assert!(result.contains(2));
-        assert!(result.contains(3));
-    }
-
-    #[test]
-    fn test_postings_equal_list_matcher() {
-        // Test case for decomposed regex node[12] -> Equal(List(["node1", "node2"]))
-        let mut postings = Postings::default();
-
-        // Add test series with various node labels
-        postings.add_posting_for_label_value(1, "node", "node1");
-        postings.add_posting_for_label_value(2, "node", "node2");
-        postings.add_posting_for_label_value(3, "node", "node3");
-        postings.add_posting_for_label_value(4, "node", "node1");
-
-        // Create a label filter for node = ("node1", "node2")
-        // This is what parse_regex_matcher("node[12]", true) should now produce
-        let lf = LabelFilter {
-            label: "node".to_string(),
-            matcher: PredicateMatch::Equal(PredicateValue::from(vec![
-                "node1".to_string(),
-                "node2".to_string(),
-            ])),
-        };
-
-        // Query using this filter
-        let result = postings.postings_for_label_filters(&[lf]).unwrap();
-        let result = result.into_owned();
-
-        // Should match series 1, 2, and 4 (all with node1 or node2)
-        assert_eq!(result.cardinality(), 3);
-        assert!(result.contains(1));
-        assert!(result.contains(2));
-        assert!(result.contains(4));
-        assert!(!result.contains(3)); // node3 should not be matched
-    }
-
-    #[test]
-    fn test_match_none_filter_returns_empty_postings() {
-        let mut postings = Postings::default();
-        postings.add_posting_for_label_value(1, "instance", "server-1");
-        postings.add_posting_for_label_value(2, "instance", "client-1");
-
-        let filter = LabelFilter {
-            label: "instance".to_string(),
-            matcher: PredicateMatch::MatchNone,
-        };
-
-        let result = postings.postings_for_label_filters(&[filter]).unwrap();
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn test_not_starts_with_filter_returns_non_matching_label_values() {
-        let mut postings = Postings::default();
-        postings.add_posting_for_label_value(1, "instance", "server-1");
-        postings.add_posting_for_label_value(2, "instance", "client-1");
-        postings.add_posting_for_label_value(3, "instance", "server-2");
-        postings.add_posting_for_label_value(4, "other", "value");
-
-        let filter = LabelFilter {
-            label: "instance".to_string(),
-            matcher: PredicateMatch::NotStartsWith(PredicateValue::String("server".to_string())),
-        };
-
-        let result = postings.postings_for_filter(&filter).unwrap();
-        // Prometheus-compatible negative matchers include series that do not have the label.
-        assert_eq!(result.cardinality(), 2);
-        assert!(result.contains(2));
-        assert!(result.contains(4));
-        assert!(!result.contains(1));
-        assert!(!result.contains(3));
-    }
-
-    #[test]
-    fn test_not_starts_with_empty_prefix_returns_only_series_without_label() {
-        let mut postings = Postings::default();
-        postings.add_posting_for_label_value(1, "instance", "server-1");
-        postings.add_posting_for_label_value(2, "instance", "client-1");
-        postings.all_postings.add(4);
-
-        let err = LabelFilter::create(MatchOp::NotStartsWith, "instance", "")
-            .expect_err("empty prefixes should be rejected");
-        assert!(
-            err.to_string()
-                .contains("starts with matcher does not allow empty values")
-        );
-    }
-
-    #[test]
-    fn test_starts_with_filter_with_value_list_uses_any_match_semantics() {
-        let mut postings = Postings::default();
-        postings.add_posting_for_label_value(1, "instance", "server-1");
-        postings.add_posting_for_label_value(2, "instance", "client-1");
-        postings.add_posting_for_label_value(3, "instance", "proxy-1");
-
-        let filter = LabelFilter {
-            label: "instance".to_string(),
-            matcher: PredicateMatch::StartsWith(PredicateValue::from(vec![
-                "server".to_string(),
-                "client".to_string(),
-            ])),
-        };
-
-        let result = postings.postings_for_filter(&filter).unwrap();
-        assert_eq!(result.cardinality(), 2);
-        assert!(result.contains(1));
-        assert!(result.contains(2));
-        assert!(!result.contains(3));
-    }
-
-    #[test]
-    fn test_not_starts_with_filter_with_value_list_requires_all_absent() {
-        let mut postings = Postings::default();
-        postings.add_posting_for_label_value(1, "instance", "server-1");
-        postings.add_posting_for_label_value(2, "instance", "client-1");
-        postings.add_posting_for_label_value(3, "instance", "proxy-1");
-        postings.all_postings.add(4);
-
-        let filter = LabelFilter {
-            label: "instance".to_string(),
-            matcher: PredicateMatch::NotStartsWith(PredicateValue::from(vec![
-                "server".to_string(),
-                "client".to_string(),
-            ])),
-        };
-
-        let result = postings.postings_for_label_filters(&[filter]).unwrap();
-        let result = result.into_owned();
-        assert_eq!(result.cardinality(), 2);
-        assert!(result.contains(3));
-        assert!(result.contains(4));
-        assert!(!result.contains(1));
-        assert!(!result.contains(2));
-    }
-
-    #[test]
-    fn test_contains_filter_returns_matching_label_values() {
-        let mut postings = Postings::default();
-        postings.add_posting_for_label_value(1, "instance", "server-1");
-        postings.add_posting_for_label_value(2, "instance", "client-1");
-        postings.add_posting_for_label_value(3, "instance", "web-server-2");
-        postings.add_posting_for_label_value(4, "other", "value");
-
-        let filter = LabelFilter {
-            label: "instance".to_string(),
-            matcher: PredicateMatch::Contains(PredicateValue::String("server".to_string())),
-        };
-
-        let result = postings.postings_for_filter(&filter).unwrap();
-
-        assert_eq!(result.cardinality(), 2);
-        assert!(result.contains(1));
-        assert!(result.contains(3));
-        assert!(!result.contains(2));
-        assert!(!result.contains(4));
-    }
-
-    #[test]
-    fn test_not_contains_filter_returns_non_matching_and_missing_labels() {
-        let mut postings = Postings::default();
-        postings.add_posting_for_label_value(1, "instance", "server-1");
-        postings.add_posting_for_label_value(2, "instance", "client-1");
-        postings.add_posting_for_label_value(3, "instance", "web-server-2");
-        postings.add_posting_for_label_value(4, "other", "value");
-        postings.all_postings.add(4);
-
-        let filter = LabelFilter {
-            label: "instance".to_string(),
-            matcher: PredicateMatch::NotContains(PredicateValue::String("server".to_string())),
-        };
-
-        let result = postings.postings_for_label_filters(&[filter]).unwrap();
-        let result = result.into_owned();
-
-        assert_eq!(result.cardinality(), 2);
-        assert!(result.contains(2));
-        assert!(result.contains(4));
-        assert!(!result.contains(1));
-        assert!(!result.contains(3));
-    }
-
-    #[test]
-    fn test_contains_filter_with_value_list_uses_any_match_semantics() {
-        let mut postings = Postings::default();
-        postings.add_posting_for_label_value(1, "instance", "server-1");
-        postings.add_posting_for_label_value(2, "instance", "client-1");
-        postings.add_posting_for_label_value(3, "instance", "proxy-1");
-
-        let filter = LabelFilter {
-            label: "instance".to_string(),
-            matcher: PredicateMatch::Contains(PredicateValue::from(vec![
-                "server".to_string(),
-                "client".to_string(),
-            ])),
-        };
-
-        let result = postings.postings_for_filter(&filter).unwrap();
-        assert_eq!(result.cardinality(), 2);
-        assert!(result.contains(1));
-        assert!(result.contains(2));
-        assert!(!result.contains(3));
-    }
-
-    #[test]
-    fn test_not_contains_filter_with_value_list_requires_all_absent() {
-        let mut postings = Postings::default();
-        postings.add_posting_for_label_value(1, "instance", "server-1");
-        postings.add_posting_for_label_value(2, "instance", "client-1");
-        postings.add_posting_for_label_value(3, "instance", "proxy-1");
-        postings.all_postings.add(4);
-
-        let filter = LabelFilter {
-            label: "instance".to_string(),
-            matcher: PredicateMatch::NotContains(PredicateValue::from(vec![
-                "server".to_string(),
-                "client".to_string(),
-            ])),
-        };
-
-        let result = postings.postings_for_label_filters(&[filter]).unwrap();
-        let result = result.into_owned();
-        assert_eq!(result.cardinality(), 2);
-        assert!(result.contains(3));
-        assert!(result.contains(4));
-        assert!(!result.contains(1));
-        assert!(!result.contains(2));
-    }
-
-    #[test]
-    fn test_postings_for_label_value_exceeds_stack_size() {
-        let mut postings = Postings::default();
-
-        // The STACK_SIZE in KeyBuffer is 64 bytes
-        // We need label_name + "=" + value + "\0" to exceed 64 bytes
-        // Let's create a label name and value that together exceed this
-
-        // Create a label name of 30 characters
-        let label_name = "very_long_label_name_here_1234";
-
-        // Create a value of 40 characters, so the total length is:
-        // 30 (label) + 1 (=) + 40 (value) + 1 (\0) = 72 bytes > 64
-        let value = "this_is_a_very_long_value_string_12345";
-
-        // Verify our assumption about the length
-        let total_len = label_name.len() + 1 + value.len() + 1; // +1 for '=', +1 for '\0'
-        assert!(
-            total_len > 64,
-            "Test setup error: combined length should exceed STACK_SIZE"
-        );
-
-        // Add a posting with this long label-value pair
-        postings.add_posting_for_label_value(1, label_name, value);
-        postings.add_posting_for_label_value(2, label_name, value);
-        postings.add_posting_for_label_value(3, label_name, "short");
-
-        // Test that we can retrieve postings for the long label-value pair
-        let result = postings.postings_for_label_value(label_name, value);
-
-        // Verify the result contains the correct series IDs
-        assert_eq!(result.cardinality(), 2);
-        assert!(result.contains(1));
-        assert!(result.contains(2));
-        assert!(!result.contains(3));
-
-        // Test that we can also retrieve the short value
-        let result_short = postings.postings_for_label_value(label_name, "short");
-        assert_eq!(result_short.cardinality(), 1);
-        assert!(result_short.contains(3));
-    }
-
-    #[test]
-    fn test_postings_for_all_label_values_exceeds_stack_size() {
-        let mut postings = Postings::default();
-
-        // The STACK_SIZE in KeyBuffer is 64 bytes
-        // KeyBuffer::for_prefix creates: label_name + "="
-        // We need label_name + "=" to exceed 64 bytes
-
-        // Create a label name of 70 characters to exceed the stack size
-        let label_name = "very_long_label_name_here_that_definitely_exceeds_the_stack_buffer_size";
-
-        // Verify our assumption about the length
-        let prefix_len = label_name.len() + 1; // +1 for '='
-        assert!(
-            prefix_len > 64,
-            "Test setup error: prefix length should exceed STACK_SIZE"
-        );
-
-        // Add multiple postings with different values for the same long label name
-        postings.add_posting_for_label_value(1, label_name, "value1");
-        postings.add_posting_for_label_value(2, label_name, "value2");
-        postings.add_posting_for_label_value(3, label_name, "value3");
-        postings.add_posting_for_label_value(4, "short_label", "value1");
-
-        // Test that we can retrieve all postings for the long label name
-        let result = postings.postings_for_all_label_values(label_name);
-
-        // Verify the result contains all series IDs with the long label name
-        assert_eq!(result.cardinality(), 3);
-        assert!(result.contains(1));
-        assert!(result.contains(2));
-        assert!(result.contains(3));
-        assert!(!result.contains(4)); // This has a different label
-
-        // Also test that we can retrieve postings for the short label
-        let result_short = postings.postings_for_all_label_values("short_label");
-        assert_eq!(result_short.cardinality(), 1);
-        assert!(result_short.contains(4));
-    }
-
-    #[test]
-    fn test_memory_postings_set_timeseries_key() {
-        let mut postings = Postings::default();
-
-        postings.set_timeseries_key(1, b"key1");
-        postings.set_timeseries_key(2, b"key2");
-
-        assert_eq!(
-            postings.get_key_by_id(1),
-            Some(&b"key1".to_vec().into_boxed_slice())
-        );
-        assert_eq!(
-            postings.get_key_by_id(2),
-            Some(&b"key2".to_vec().into_boxed_slice())
-        );
-        assert_eq!(postings.get_key_by_id(3), None);
-    }
-
-    #[test]
-    fn test_memory_postings_remove_timeseries() {
-        let mut postings = Postings::default();
-        let mut series = TimeSeries::new();
-        series.id = 1;
-        series.labels = MetricName::new(&[
-            Label::new("label1", "value1"),
-            Label::new("label2", "value2"),
-        ]);
-
-        postings.add_posting_for_label_value(1, "label1", "value1");
-        postings.add_posting_for_label_value(1, "label2", "value2");
-
-        postings.remove_timeseries(&series);
-
-        assert!(
-            postings
-                .postings_for_label_value("label1", "value1")
-                .is_empty()
-        );
-        assert!(
-            postings
-                .postings_for_label_value("label2", "value2")
-                .is_empty()
-        );
-        assert!(postings.all_postings.is_empty());
-    }
-
-    #[test]
-    fn test_remove_stale_ids_across_multiple_batches() {
-        // Regression test: draining stale ids must visit *every* label key, not just those sharing
-        // a prefix with the resume cursor. Previously the resume used the cursor as a prefix, which
-        // abandoned later keys and cleared `stale_ids` while their bitmaps were still dirty.
-        let mut postings = Postings::default();
-        let n: u64 = 250;
-        for i in 0..n {
-            postings.add_posting_for_label_value(i as SeriesRef, "host", &format!("h{i:04}"));
-            postings.add_posting_for_label_value(i as SeriesRef, "job", "web");
-        }
-
-        // Mark the even ids stale.
-        let stale: Vec<SeriesRef> = (0..n).step_by(2).map(|i| i as SeriesRef).collect();
-        postings.mark_ids_as_stale(&stale);
-        assert!(postings.has_stale_ids());
-
-        // Drain in small batches, like the background maintenance task.
-        let mut cursor = None;
-        while let Some(next) = postings.remove_stale_ids(cursor.take(), 16) {
-            cursor = Some(next);
-        }
-
-        assert!(
-            !postings.has_stale_ids(),
-            "stale ids should be fully drained"
-        );
-
-        // The shared `job=web` bitmap sorts after all the `host=*` keys, so with the old
-        // prefix-based resume it was never cleaned. Inspect the raw bitmap directly (queries mask
-        // stale ids regardless of whether the bitmap was physically cleaned).
-        let key = IndexKey::for_label_value("job", "web");
-        let bmp = postings
-            .label_index
-            .get::<IndexKey>(&key)
-            .expect("job=web posting list should exist");
-        for id in &stale {
-            assert!(
-                !bmp.contains(*id),
-                "stale id {id} should have been removed from the bitmap"
-            );
-        }
-        // Odd (non-stale) ids remain.
-        assert!(bmp.contains(1));
-        assert!(bmp.contains(3));
-    }
-
-    #[test]
-    fn test_memory_postings_postings_by_labels() {
-        let mut postings = Postings::default();
-
-        postings.add_posting_for_label_value(1, "label1", "value1");
-        postings.add_posting_for_label_value(1, "label2", "value2");
-        postings.add_posting_for_label_value(2, "label1", "value1");
-        postings.add_posting_for_label_value(2, "label2", "value3");
-
-        let labels = vec![
-            Label::new("label1", "value1"),
-            Label::new("label2", "value2"),
-        ];
-        let result = postings.postings_by_labels(&labels);
-
-        assert_eq!(result.cardinality(), 1);
-        assert!(result.contains(1));
-    }
-
-    #[test]
-    fn test_decompose_vs_fullmatch_server_wildcard() {
-        // Test to verify that decomposed regex matching for "server.*" works correctly
-        use regex::Regex;
-
-        // Pattern: "server.*" decomposes to prefix "server" + remainder ".*" (compiled as ^.*$)
-        let re_decomposed = Regex::new("^.*$").unwrap();
-
-        // Full pattern: "server.*" (for reference)
-        let re_full = Regex::new("^server.*$").unwrap();
-
-        let test_values = vec!["server1", "server", "serverx", "serverabc"];
-
-        for val in test_values {
-            // Check full regex
-            let full_match = re_full.is_match(val);
-
-            // Check decomposed: if starts with "server", check if remainder matches ^.*$
-            let decomposed_match = if let Some(remainder) = val.strip_prefix("server") {
-                re_decomposed.is_match(remainder)
-            } else {
-                false
-            };
-
-            assert_eq!(
-                full_match, decomposed_match,
-                "Mismatch for value '{}': full={}, decomposed={}",
-                val, full_match, decomposed_match
-            );
-
-            // debug printing removed
-        }
-    }
-}
